fix(sync-actions): reference corresponding images on both old and new variants
affects: @commercetools/sync-actions

if the order of images in the array should change, the module references both images by index, but
they were diffed uniquely by url. So now we are referencing corresponding images with the url, if a
url on old is found in the new variant then that means it's the same image.