fix(sdk-middleware-http): retry on http response with status 503 (#913)
#### Summary

When the server responds with a `503` status and `retryEnabled` is enabled, no retry is performed. But documentation describes that a retry will be performed in this case. 

#### Description

We created a test in our `commercetools-importer` project that returned a response with status `503` and expected that this will trigger a retry. But this didn't happened and so we took a look at the current implementation. And the code looked suspicious because according to the fetch documentation, only network errors are thrown as error and `5xx` are returned as an error response.

#### Todo

- Tests
  - [ ] Unit
  - [ ] Integration
  - [ ] Acceptance
- [ ] Documentation