chore: add keyword for automation (#1085)
#### Summary

This PR adds the `cli` keyword to packages used by IMPEX to enable automation of the "CLI Overview" page on the docs site. See DOCS-44 for more information.

It also fixes a few grammar issues in the descriptions, since those are getting pulled in too ;)

<!-- Provide a short summary of your changes -->

#### Description

<!-- Describe the changes in this PR here and provide some context -->

#### Todo

I'm really not sure what kinds of testing this will need? :)

- Tests
  - [x] Unit
  - [x] Integration
  - [x] Acceptance
- [ / ] Documentation
  <!-- Two persons should review a PR, don't forget to assign them. -->