fix(sdk-middleware-http) Remove dependence on content in ok result
It's not uncommon to use return `Accepted()` or return `Ok()` in asp core and even though specifying the content type the return will be empty making json() throw. The content length header is stripped in the proxy so the simplest solution for handling a common server solution for empty 202 and 200 seems to be to catch the exception and continue with an empty body.