test: add test for long text performance
This test fails in the current state, as jsondiffpatch
uses the diff-match-patch library for long text diffing.
This text diff is very slow, but produces a nice delta that
can be used to see the difference between two text strings.
However, we don't actually care about identifying the specific
differences between two strings - we just care if they changed.
The regular text diffing strategy is MUCH faster, and performance
is what we want to prioritize.

- https://github.com/benjamine/jsondiffpatch/blob/master/docs/deltas.md#text-diffs
- https://github.com/google/diff-match-patch